pyproject.toml
[project]
name = "windguru-archive-browser"
version = "1.0.0"
description = "A powerful command-line tool for fetching and visualizing historical wind data from Windguru"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "fgeck", email = "fgeck@users.noreply.github.com" }
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.14.0",
    "lxml>=6.0.0",
    "pandas>=2.3.0",
    "numpy>=1.26.0",
    "plotly>=6.5.0",
    "kaleido>=1.2.0",
    "python-dateutil>=2.8.2",
    "keyring>=25.0.0",
    "textual>=1.0.0",
]

[project.scripts]
windguru = "windguru:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = false
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = false
strict_equality = false
